+++ config/urls.py
@@ -15,6 +15,7 @@ urlpatterns = [
     path("accounts/", include("allauth.urls")),
     # Django controlled vocabulary
     path("vocabularies/", include("controlled_vocabulary.urls")),
+    path("resources/", include("radical_translations.core.urls")),
     # Wagtail
     path("cms/", include(wagtailadmin_urls)),
     path("documents/", include(wagtaildocs_urls)),